TST: separate out pd.crosstab tests from test_pivot (#32536)
